Rename Assignment components
